[frontend] Implement checkboxes / massive operations in Observations => Infrastructures (#issue/3505)
