--- opencti-platform/opencti-front/src/private/components/observations/infrastructures/InfrastructureEdition.jsx
@@ -1,114 +1,0 @@
-import React, { Component } from 'react';
-import * as PropTypes from 'prop-types';
-import { compose } from 'ramda';
-import withStyles from '@mui/styles/withStyles';
-import Drawer from '@mui/material/Drawer';
-import Fab from '@mui/material/Fab';
-import { Edit } from '@mui/icons-material';
-import { graphql } from 'react-relay';
-import { commitMutation, QueryRenderer } from '../../../../relay/environment';
-import inject18n from '../../../../components/i18n';
-import InfrastructureEditionContainer from './InfrastructureEditionContainer';
-import { infrastructureEditionOverviewFocus } from './InfrastructureEditionOverview';
-import Loader from '../../../../components/Loader';
-
-const styles = (theme) => ({
-  editButton: {
-    position: 'fixed',
-    bottom: 30,
-    right: 30,
-  },
-  drawerPaper: {
-    minHeight: '100vh',
-    width: '50%',
-    position: 'fixed',
-    overflow: 'auto',
-    transition: theme.transitions.create('width', {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.enteringScreen,
-    }),
-    padding: 0,
-  },
-});
-
-export const infrastructureEditionQuery = graphql`
-  query InfrastructureEditionContainerQuery($id: String!) {
-    infrastructure(id: $id) {
-      ...InfrastructureEditionContainer_infrastructure
-    }
-  }
-`;
-
-class InfrastructureEdition extends Component {
-  constructor(props) {
-    super(props);
-    this.state = { open: false };
-  }
-
-  handleOpen() {
-    this.setState({ open: true });
-  }
-
-  handleClose() {
-    commitMutation({
-      mutation: infrastructureEditionOverviewFocus,
-      variables: {
-        id: this.props.infrastructureId,
-        input: { focusOn: '' },
-      },
-    });
-    this.setState({ open: false });
-  }
-
-  render() {
-    const { classes, infrastructureId } = this.props;
-    return (
-      <div>
-        <Fab
-          onClick={this.handleOpen.bind(this)}
-          color="secondary"
-          aria-label="Edit"
-          className={classes.editButton}
-        >
-          <Edit />
-        </Fab>
-        <Drawer
-          open={this.state.open}
-          anchor="right"
-          sx={{ zIndex: 1202 }}
-          elevation={1}
-          classes={{ paper: classes.drawerPaper }}
-          onClose={this.handleClose.bind(this)}
-        >
-          <QueryRenderer
-            query={infrastructureEditionQuery}
-            variables={{ id: infrastructureId }}
-            render={({ props }) => {
-              if (props) {
-                return (
-                  <InfrastructureEditionContainer
-                    infrastructure={props.infrastructure}
-                    handleClose={this.handleClose.bind(this)}
-                  />
-                );
-              }
-              return <Loader variant="inElement" />;
-            }}
-          />
-        </Drawer>
-      </div>
-    );
-  }
-}
-
-InfrastructureEdition.propTypes = {
-  infrastructureId: PropTypes.string,
-  classes: PropTypes.object,
-  theme: PropTypes.object,
-  t: PropTypes.func,
-};
-
-export default compose(
-  inject18n,
-  withStyles(styles, { withTheme: true }),
-)(InfrastructureEdition);
