+++ opencti-platform/opencti-front/src/private/components/observations/infrastructures/InfrastructurePopover.tsx
@@ -1,0 +1,174 @@
+import React, { useState } from 'react';
+import Menu from '@mui/material/Menu';
+import MenuItem from '@mui/material/MenuItem';
+import Button from '@mui/material/Button';
+import IconButton from '@mui/material/IconButton';
+import Drawer from '@mui/material/Drawer';
+import Dialog from '@mui/material/Dialog';
+import DialogActions from '@mui/material/DialogActions';
+import DialogContent from '@mui/material/DialogContent';
+import DialogContentText from '@mui/material/DialogContentText';
+import MoreVert from '@mui/icons-material/MoreVert';
+import { graphql, useMutation } from 'react-relay';
+import makeStyles from '@mui/styles/makeStyles';
+import { useNavigate } from 'react-router-dom-v5-compat';
+import { useFormatter } from '../../../../components/i18n';
+import Loader, { LoaderVariant } from '../../../../components/Loader';
+import Security from '../../../../utils/Security';
+import { KNOWLEDGE_KNUPDATE_KNDELETE } from '../../../../utils/hooks/useGranted';
+import { Theme } from '../../../../components/Theme';
+import { InfrastructureEditionContainerQuery } from './__generated__/InfrastructureEditionContainerQuery.graphql';
+import useQueryLoading from '../../../../utils/hooks/useQueryLoading';
+import InfrastructureEditionContainer, { infrastructureEditionContainerQuery } from './InfrastructureEditionContainer';
+import Transition from '../../../../components/Transition';
+
+const useStyles = makeStyles<Theme>((theme) => ({
+  container: {
+    margin: 0,
+  },
+  drawerPaper: {
+    minHeight: '100vh',
+    width: '50%',
+    position: 'fixed',
+    overflow: 'auto',
+    transition: theme.transitions.create('width', {
+      easing: theme.transitions.easing.sharp,
+      duration: theme.transitions.duration.enteringScreen,
+    }),
+    padding: 0,
+  },
+}));
+
+const InfrastructurePopoverDeletionMutation = graphql`
+  mutation InfrastructurePopoverDeletionMutation($id: ID!) {
+    infrastructureEdit(id: $id) {
+      delete
+    }
+  }
+`;
+
+const InfrastructurePopover = ({ id }: { id: string }) => {
+  const { t } = useFormatter();
+  const classes = useStyles();
+  const navigate = useNavigate();
+
+  const [anchorEl, setAnchorEl] = useState<HTMLElement | null>(null);
+  const [displayDelete, setDisplayDelete] = useState<boolean>(false);
+  const [displayEdit, setDisplayEdit] = useState<boolean>(false);
+  const [deleting, setDeleting] = useState<boolean>(false);
+
+  const [commit] = useMutation(InfrastructurePopoverDeletionMutation);
+  const queryRef = useQueryLoading<InfrastructureEditionContainerQuery>(
+    infrastructureEditionContainerQuery,
+    { id },
+  );
+
+  const handleOpen = (event: React.MouseEvent<HTMLElement>) => {
+    setAnchorEl(event.currentTarget);
+  };
+
+  const handleClose = () => setAnchorEl(null);
+
+  const handleOpenDelete = () => {
+    setDisplayDelete(true);
+    handleClose();
+  };
+
+  const handleCloseDelete = () => setDisplayDelete(false);
+
+  const submitDelete = () => {
+    setDeleting(true);
+    commit({
+      variables: {
+        id,
+      },
+      onCompleted: () => {
+        setDeleting(false);
+        handleClose();
+        navigate('/dashboard/observations/infrastructures');
+      },
+    });
+  };
+
+  const handleOpenEdit = () => {
+    setDisplayEdit(true);
+    handleClose();
+  };
+
+  const handleCloseEdit = () => setDisplayEdit(false);
+  return (
+    <div className={classes.container}>
+      <IconButton
+        onClick={handleOpen}
+        aria-haspopup="true"
+        style={{ marginTop: 3 }}
+        size="large"
+      >
+        <MoreVert />
+      </IconButton>
+      <Menu
+        anchorEl={anchorEl}
+        open={Boolean(anchorEl)}
+        onClose={handleClose}
+      >
+        <MenuItem onClick={handleOpenEdit}>
+          {t('Update')}
+        </MenuItem>
+        <Security needs={[KNOWLEDGE_KNUPDATE_KNDELETE]}>
+          <MenuItem onClick={handleOpenDelete}>
+            {t('Delete')}
+          </MenuItem>
+        </Security>
+      </Menu>
+      <Dialog
+        open={displayDelete}
+        PaperProps={{ elevation: 1 }}
+        keepMounted={true}
+        TransitionComponent={Transition}
+        onClose={handleCloseDelete}
+      >
+        <DialogContent>
+          <DialogContentText>
+            {t('Do you want to delete this intrusion set?')}
+          </DialogContentText>
+        </DialogContent>
+        <DialogActions>
+          <Button
+            onClick={handleCloseDelete}
+            disabled={deleting}
+          >
+            {t('Cancel')}
+          </Button>
+          <Button
+            color="secondary"
+            onClick={submitDelete}
+            disabled={deleting}
+          >
+            {t('Delete')}
+          </Button>
+        </DialogActions>
+      </Dialog>
+      <Drawer
+        open={displayEdit}
+        anchor="right"
+        sx={{ zIndex: 1202 }}
+        elevation={1}
+        classes={{ paper: classes.drawerPaper }}
+        onClose={handleCloseEdit}
+      >
+        {queryRef && (
+          <React.Suspense
+            fallback={<Loader variant={LoaderVariant.inElement} />}
+          >
+            <InfrastructureEditionContainer
+              queryRef={queryRef}
+              handleClose={handleClose}
+            />
+          </React.Suspense>
+        )}
+      </Drawer>
+    </div>
+  );
+};
+
+export default InfrastructurePopover;
