+++ opencti-platform/opencti-front/src/private/components/observations/infrastructures/Root.tsx
@@ -3,31 +3,29 @@
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-nocheck
 import React, { useMemo } from 'react';
-import { Route, Redirect, useParams } from 'react-router-dom';
-import { graphql, useSubscription } from 'react-relay';
+import {Route, Redirect, useParams, Switch} from 'react-router-dom';
+import {graphql, usePreloadedQuery, useSubscription} from 'react-relay';
 import { GraphQLSubscriptionConfig } from 'relay-runtime';
-import {
-  QueryRenderer,
-} from '../../../../relay/environment';
 import TopBar from '../../nav/TopBar';
-import Infrastructure from './Infrastructure';
 import InfrastructureKnowledge from './InfrastructureKnowledge';
 import StixDomainObjectHeader from '../../common/stix_domain_objects/StixDomainObjectHeader';
 import FileManager from '../../common/files/FileManager';
 import InfrastructurePopover from './InfrastructurePopover';
-import Loader from '../../../../components/Loader';
+import Loader, {LoaderVariant} from '../../../../components/Loader';
 import StixCoreObjectHistory from '../../common/stix_core_objects/StixCoreObjectHistory';
 import StixCoreObjectOrStixCoreRelationshipContainers from '../../common/containers/StixCoreObjectOrStixCoreRelationshipContainers';
 import ErrorNotFound from '../../../../components/ErrorNotFound';
 import { RootInfrastructureSubscription } from './__generated__/RootInfrastructureSubscription.graphql';
-import { RootInfrastructureQuery$data } from './__generated__/RootInfrastructureQuery.graphql';
+import {RootInfrastructureQuery} from './__generated__/RootInfrastructureQuery.graphql';
+import Infrastructure from './Infrastructure';
+import useQueryLoading from "../../../../utils/hooks/useQueryLoading";
 
 const subscription = graphql`
   subscription RootInfrastructureSubscription($id: ID!) {
     stixDomainObject(id: $id) {
       ... on Infrastructure {
         ...Infrastructure_infrastructure
-        ...InfrastructureEditionContainer_infrastructure
+        ...InfrastructureEditionOverview_infrastructure
       }
       ...FileImportViewer_entity
       ...FileExportViewer_entity
@@ -63,8 +61,7 @@ const infrastructureQuery = graphql`
   }
 `;
 
-const RootInfrastructureComponent = () => {
-  const { infrastructureId } = useParams() as { infrastructureId: string };
+const RootInfrastructureComponent = ({ queryRef, infrastructureId }) => {
   const subConfig = useMemo<
   GraphQLSubscriptionConfig<RootInfrastructureSubscription>
   >(
@@ -75,109 +72,107 @@ const RootInfrastructureComponent = () => {
     [infrastructureId],
   );
   useSubscription(subConfig);
+  const data = usePreloadedQuery(infrastructureQuery, queryRef);
+  const { infrastructure, connectorsForImport, connectorsForExport } = data;
   return (
-    <div>
-      <TopBar />
-      <QueryRenderer
-        query={infrastructureQuery}
-        variables={{ id: infrastructureId }}
-        render={({ props }: { props: RootInfrastructureQuery$data }) => {
-          if (props) {
-            if (props.infrastructure) {
-              return (
-                <div>
-                  <Route
-                    exact
-                    path="/dashboard/observations/infrastructures/:infrastructureId"
-                    render={(routeProps) => (
-                      <Infrastructure
-                        {...routeProps}
-                        infrastructure={props.infrastructure}
-                      />
-                    )}
-                  />
-                  <Route
-                    exact
-                    path="/dashboard/observations/infrastructures/:infrastructureId/knowledge"
-                    render={() => (
-                      <Redirect
-                        to={`/dashboard/observations/infrastructures/${infrastructureId}/knowledge/overview`}
-                      />
-                    )}
-                  />
-                  <Route
-                    path="/dashboard/observations/infrastructures/:infrastructureId/knowledge"
-                    render={(routeProps) => (
-                      <InfrastructureKnowledge
-                        {...routeProps}
-                        infrastructure={props.infrastructure}
-                      />
-                    )}
-                  />
-                  <Route
-                    exact
-                    path="/dashboard/observations/infrastructures/:infrastructureId/analyses"
-                    render={(routeProps) => (
-                      <React.Fragment>
-                        <StixDomainObjectHeader
-                          entityType={'Infrastructure'}
-                          stixDomainObject={props.infrastructure}
-                          PopoverComponent={<InfrastructurePopover />}
-                        />
-                        <StixCoreObjectOrStixCoreRelationshipContainers
-                          {...routeProps}
-                          stixDomainObjectOrStixCoreRelationship={
-                            props.infrastructure
-                          }
-                        />
-                      </React.Fragment>
-                    )}
-                  />
-                  <Route
-                    exact
-                    path="/dashboard/observations/infrastructures/:infrastructureId/files"
-                    render={(routeProps) => (
-                      <React.Fragment>
-                        <StixDomainObjectHeader
-                          stixDomainObject={props.infrastructure}
-                          PopoverComponent={<InfrastructurePopover />}
-                        />
-                        <FileManager
-                          {...routeProps}
-                          id={infrastructureId}
-                          connectorsImport={props.connectorsForImport}
-                          connectorsExport={props.connectorsForExport}
-                          entity={props.infrastructure}
-                        />
-                      </React.Fragment>
-                    )}
-                  />
-                  <Route
-                    exact
-                    path="/dashboard/observations/infrastructures/:infrastructureId/history"
-                    render={(routeProps) => (
-                      <React.Fragment>
-                        <StixDomainObjectHeader
-                          stixDomainObject={props.infrastructure}
-                          PopoverComponent={<InfrastructurePopover />}
-                        />
-                        <StixCoreObjectHistory
-                          {...routeProps}
-                          stixCoreObjectId={infrastructureId}
-                        />
-                      </React.Fragment>
-                    )}
-                  />
-                </div>
-              );
-            }
-            return <ErrorNotFound />;
-          }
-          return <Loader />;
-        }}
-      />
-    </div>
+    <>
+      {infrastructure ? (
+        <Switch>
+          <Route
+            exact
+            path="/dashboard/observations/infrastructures/:infrastructureId"
+            render={() => <Infrastructure data={infrastructure} />}
+          />
+          <Route
+            exact
+            path="/dashboard/observations/infrastructures/:infrastructureId/knowledge"
+            render={() => (
+              <Redirect
+                to={`/dashboard/observations/infrastructures/${infrastructureId}/knowledge/overview`}
+              />
+            )}
+          />
+          <Route
+            path="/dashboard/observations/infrastructures/:infrastructureId/knowledge"
+            render={() => (
+              <InfrastructureKnowledge infrastructure={infrastructure} />
+            )}
+          />
+          <Route
+            exact
+            path="/dashboard/observations/infrastructures/:infrastructureId/analyses"
+            render={(routeProps) => (
+              <React.Fragment>
+                <StixDomainObjectHeader
+                  entityType={'Infrastructure'}
+                  stixDomainObject={infrastructure}
+                  PopoverComponent={InfrastructurePopover}
+                />
+                <StixCoreObjectOrStixCoreRelationshipContainers
+                  {...routeProps}
+                  stixDomainObjectOrStixCoreRelationship={infrastructure}
+                />
+              </React.Fragment>
+            )}
+          />
+          <Route
+            exact
+            path="/dashboard/observations/infrastructures/:infrastructureId/files"
+            render={(routeProps) => (
+              <React.Fragment>
+                <StixDomainObjectHeader
+                  stixDomainObject={infrastructure}
+                  PopoverComponent={InfrastructurePopover}
+                />
+                <FileManager
+                  {...routeProps}
+                  id={infrastructureId}
+                  connectorsImport={connectorsForImport}
+                  connectorsExport={connectorsForExport}
+                  entity={infrastructure}
+                />
+              </React.Fragment>
+            )}
+          />
+          <Route
+            exact
+            path="/dashboard/observations/infrastructures/:infrastructureId/history"
+            render={(routeProps) => (
+              <React.Fragment>
+                <StixDomainObjectHeader
+                  entityType={'Infrastructure'}
+                  disableSharing={true}
+                  stixDomainObject={infrastructure}
+                  PopoverComponent={InfrastructurePopover}
+                />
+                <StixCoreObjectHistory
+                  {...routeProps}
+                  stixCoreObjectId={infrastructureId}
+                />
+              </React.Fragment>
+            )}
+          />
+        </Switch>
+      ) : (
+        <ErrorNotFound />
+      )}
+    </>
   );
 };
 
-export default RootInfrastructureComponent;
+const RootInfrastructure = () => {
+  const { infrastructureId } = useParams() as { infrastructureId: string };
+  const queryRef = useQueryLoading<RootInfrastructureQuery>(infrastructureQuery, { id: infrastructureId });
+  return (
+    <>
+      <TopBar/>
+      {queryRef && (
+        <React.Suspense fallback={<Loader variant={LoaderVariant.container} />}>
+          <RootInfrastructureComponent queryRef={queryRef} infrastructureId={infrastructureId} />
+        </React.Suspense>
+        )}
+    </>
+  )
+}
+
+export default RootInfrastructure;
